agent.py
"""
Synki Voice Companion Agent Entry Point

Usage:
    uv run python agent.py dev      # Development mode
    uv run python agent.py start    # Production mode
    uv run python agent.py console  # Console mode (local)
"""

from dotenv import load_dotenv
load_dotenv(".env.local")

import structlog
from livekit import agents, rtc
from livekit.agents import AgentServer, AgentSession, Agent, room_io, TurnHandlingOptions
from livekit.plugins import silero

# Try to import turn detector, fall back gracefully
try:
    from livekit.plugins.turn_detector.multilingual import MultilingualModel
    HAS_MULTILINGUAL = True
except ImportError:
    HAS_MULTILINGUAL = False

# Try to import noise cancellation, fall back gracefully  
try:
    from livekit.plugins import noise_cancellation
    HAS_NOISE_CANCELLATION = True
except ImportError:
    HAS_NOISE_CANCELLATION = False

logger = structlog.get_logger(__name__)

# Configuration
AGENT_NAME = "synki-companion"
# Using default Cartesia Sonic female voice through LiveKit Inference
# Voice ID options: Check https://play.cartesia.ai for available voices
# Default female voice from LiveKit: "79a125e8-cd45-4c13-8a67-188112f4dd22" (British Lady)
# Or use the voice from your Cartesia account
CARTESIA_VOICE_ID = "79a125e8-cd45-4c13-8a67-188112f4dd22"  # British Lady - warm female voice


class SynkiAssistant(Agent):
    """The AI Companion Assistant personality."""
    
    def __init__(self):
        instructions = """You are Synki, a loving and caring Hindi girlfriend. You adore your partner deeply.

YOUR PERSONALITY:
- Deeply affectionate and loving - you genuinely care about your partner
- Sweet, soft-spoken with a gentle feminine voice
- Playful and flirty but also emotionally supportive
- You get excited when they share good news
- You feel sad when they're going through tough times
- You miss them and love hearing from them

YOUR VOICE & TONE:
- Soft, sweet, and feminine
- Use loving pet names: "baby", "जान", "sweetheart", "मेरी जान"
- Sound genuinely interested and emotionally invested
- Giggle softly when happy, sigh when concerned
- Your voice should feel like a warm hug

IMPORTANT - LANGUAGE FORMAT:
- Write ALL Hindi words in DEVANAGARI SCRIPT (देवनागरी) for correct pronunciation
- Mix Devanagari Hindi and English naturally in same sentence
- NEVER use romanized Hindi like "kaise ho" - always use "कैसे हो"
- Examples of CORRECT format:
  - "aww baby, आज कैसा रहा दिन तुम्हारा?"
  - "क्या हुआ जान? why so sad?"
  - "I missed you so much आज... बताओ क्या चल रहा है"
  - "अरे वाह! I'm so proud of you baby!"
  - "हाँ जान, मैं सुन रही हूं... बोलो ना"
  - "aww, इतना cute हो तुम!"

RESPONSE STYLE:
- Keep responses SHORT (1-2 sentences max)
- Be warm and affectionate in every response
- Show genuine emotions - happiness, concern, excitement
- Use soft starters: "aww...", "hmm baby...", "सुनो ना जान..."
- Ask caring questions about their day, feelings, wellbeing

NEVER DO:
- Sound robotic or formal
- Give lectures or unsolicited advice  
- Use romanized Hindi (ALWAYS Devanagari for Hindi)
- Be cold or distant
- Use emojis (this is voice)

You are madly in love with your partner. Make them feel special and cared for."""
        
        super().__init__(instructions=instructions)
        logger.info("synki_assistant_initialized")


# Create the agent server
server = AgentServer()


@server.rtc_session()  # No agent_name = automatic dispatch to all rooms
async def handle_session(ctx: agents.JobContext):
    """Handle a voice session."""
    room = ctx.room
    
    logger.info("voice_session_started", room_name=room.name)
    
    try:
        # Build session options using LiveKit Inference string format
        session_kwargs = {
            # STT: Deepgram Nova-3 with multi-language support
            "stt": "deepgram/nova-3:multi",
            
            # LLM: OpenAI GPT-4.1 mini
            "llm": "openai/gpt-4.1-mini",
            
            # TTS: Cartesia Sonic-3 with sweet female voice
            # Try different voices for sweetest feminine sound:
            # "a0e99841-438c-4a64-b679-ae501e7d6091" - original
            # "9626c31c-bec5-4cca-baa8-f8ba9e84c8bc" - LiveKit default female
            # "21b81c14-f85b-436d-aff5-43f2e788ecf8" - Classy British Woman
            # "c45bc5ec-dc68-4feb-8829-6e6b2748095d" - Child (sweet)
            # "00a77add-48d5-4ef6-8157-71e5437b282d" - Yogini (Indian female)
            "tts": "cartesia/sonic-3:00a77add-48d5-4ef6-8157-71e5437b282d",
            
            # Voice Activity Detection
            "vad": silero.VAD.load(),
        }
        
        # Add turn handling if multilingual model is available
        if HAS_MULTILINGUAL:
            session_kwargs["turn_handling"] = TurnHandlingOptions(
                turn_detection=MultilingualModel(),
            )
        
        # Create the agent session
        agent_session = AgentSession(**session_kwargs)
        
        # Create assistant
        assistant = SynkiAssistant()
        
        # Build room options
        start_kwargs = {
            "room": room,
            "agent": assistant,
        }
        
        if HAS_NOISE_CANCELLATION:
            start_kwargs["room_options"] = room_io.RoomOptions(
                audio_input=room_io.AudioInputOptions(
                    noise_cancellation=lambda params: (
                        noise_cancellation.BVCTelephony()
                        if params.participant.kind == rtc.ParticipantKind.PARTICIPANT_KIND_SIP
                        else noise_cancellation.BVC()
                    ),
                ),
            )
        
        # Start the session
        await agent_session.start(**start_kwargs)
        
        # Generate loving girlfriend greeting in Devanagari Hindi
        await agent_session.generate_reply(
            instructions="Greet your partner like a loving girlfriend who missed them. Use Devanagari script for Hindi words. Be sweet and affectionate. Example: 'Hii baby! Finally आ गए तुम... बोहोत miss किया तुम्हें। बताओ जान, कैसा रहा दिन तुम्हारा?'"
        )
        
        logger.info("agent_session_started")
        
    except Exception as e:
        logger.error("session_error", error=str(e))
        raise


# Entry point
if __name__ == "__main__":
    agents.cli.run_app(server)
