pyproject.toml
[project]
name = "synki"
version = "0.1.0"
description = "Hindi GF-style Voice Companion using LiveKit, Deepgram, OpenAI, and Cartesia"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Synki Team" }
]
keywords = ["voice-ai", "livekit", "deepgram", "openai", "cartesia", "companion"]

dependencies = [
    "livekit-agents>=0.8,<2.0",
    "livekit-plugins-deepgram>=0.6",
    "livekit-plugins-openai>=0.8",
    "livekit-plugins-cartesia>=0.4",
    "livekit-plugins-silero>=0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "mypy>=1.9",
    "black>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synki"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
